feat(backend): setup json-api interceptor and prepare docs for core folder
