+++ backend/cmd/db-tool/main.go
@@ -1,9 +1,12 @@
 package main
 
 import (
+	"db-tool/internal/api"
 	"db-tool/internal/config"
 )
 
 func main() {
 	config.Init()
+
+	api.Init()
 }
